+++ Project.swift
@@ -82,7 +82,6 @@ func targets() -> [Target] {
                 .target(name: "ProjectDescription"),
                 .target(name: "ProjectAutomation"),
                 .target(name: "TuistLoader"),
-                .target(name: "TuistInsights"),
                 .target(name: "TuistScaffold"),
                 .target(name: "TuistSigning"),
                 .target(name: "TuistDependencies"),
@@ -370,24 +369,6 @@ func targets() -> [Target] {
             name: "ProjectAutomation",
             hasTests: false,
             hasTesting: false
-        ),
-        Target.module(
-            name: "TuistInsights",
-            hasTesting: false,
-            hasIntegrationTests: true,
-            dependencies: [
-                .target(name: "TuistCore"),
-                .target(name: "TuistGraph"),
-                .target(name: "TuistSupport"),
-            ],
-            testDependencies: [
-                .target(name: "TuistSupportTesting"),
-                .target(name: "TuistGraphTesting"),
-            ],
-            integrationTestsDependencies: [
-                .target(name: "TuistSupportTesting"),
-                .target(name: "TuistGraphTesting"),
-            ]
         ),
         Target.module(
             name: "TuistSigning",