+++ Sources/ProjectDescription/Cloud.swift
@@ -11,7 +11,6 @@ public struct Cloud: Codable, Equatable {
     /// Cloud option.
     public enum Option: String, Codable, Equatable {
         case analytics
-        case insights
     }
 
     /// The base URL that points to the Cloud server